Moved styling for h1 to h6 elements to custom.css
In the process of making original.css just apply to colors.
Previously, the font-family, font-weight, line-height, and margin were specified in original.css file but that
was incorrect -- they should not change with the color but
rather should be specified in the theme itself. The h1 to h6 rules that speficy the font-family, font-weight,
line-height and margin properties were removed from original.css and added to custom.css.