Moved styling for li, ol and ul elements to custom.css
In the process of making original.css just apply to colors.
Previously these rules were inside of original.css, but they belong in custom.css. These should be constant for the theme, so it is incorrect to have them associated with a particular color scheme. The rules for the margin and padding of ol and ul lists and for the line height of li elements were removed from original.css and added under #content in custom.css.