Added color for a, a:hover and a:focus
In the process of removing original.css -- Since there is only one option for colors, original.css doesn't make much sense, so its rules are added to custom.css. Color was added to replicate the original wordpress theme. This rule specifies the color of <a> elements including the colors for a:hover and a:focus.