Removed original.css
All the elements have now been integrated into custom.css. Since
there was only one color option, it made more sense to not have a
separate file for the colors, so this file is no longer necessary.
Removed images for original color scheme

In the process of removing original.css
Since there is only one color scheme option, it should not be
a configurable setting. It does not need to be displayed in the
control panel, so the images are removed.
Removed the color-scheme option

In the process of removing original.css
Since there was only one color scheme, it doesn't make sense to
list it as an option. There will no longer be a place to make
the selection in the control panel.
Remove the @import command for original.css

In the process of removing original.css
All the color elements from original.css have been incorporated
into custom.css, so the external file is no longer needed