Footer is always affixed to the bottom of the page
Previously, the footer only appeared flush with the bottom when the
content of the page was larger than the size of the window. Otherwise,
there would be a gap between the bottom of the footer and the window.
Javascript was added to detect the times when the content is shorter than
the size of the window, and add a class with appropriate styling.