Changed rule from using navbar to dockbar class
Rule was intended to just style the dockbar menu and items, but because
the navbar class was used elsewhere (such as the wiki and message board
portlets) it was causing issues where the text was hard to read. Now it is
more specific so the other portlets will not be affected.