Created Jumbotron structure, template and example article
Previously, the first portlet in the hompage column 1 would receive this
styling, but this selector caused a lot of problems. For example, the
admin features and maximized view got the styling, and it didn't look as
intended with different layouts. Now, it is restricted to the .jumbotron
class, which is added dynamically when the jumbotron structure is used.