Bugfix: media query for the background of account and add controls
Previously, this property was in a media query, but was removed when it
was moved. It is necessary to restrict it, however, for larger widths when
the add controls in particular are displayed as icons instead of a
dropdown menu. Without the media query, the space between the icons was
being affected, which was not the intended affect.