Editing js to affix the footer and size the header
Previously these were two different "script" elements that both defined
window.onload and a window.onresize functions. However, this meant that
the scripts could not both run. A new function that is a wrapper for these
two was added to address this problem. There are still lag issues though,
and the onload script does not seem to work for the header image -- it
defaults to the CSS defined height.