Accordion toggle icons display correctly
The icon used (arrow 13_right) has the right icon and the down icon in the
same image, stacked vertically. When clicked, the position is changed to
show the menu is open. Previously there was no line-height specified, so
both image would be displayed when the menu was closed. Now the images
have the desired appearance.