Options icon displays correctly for borderless portlets
Previously this icon was missing from the top bar of borderless portlets.
The icon was provided by the liferay _unstyled theme, and there was a rule
inside of portlet.css that was supposed to place this icon in the toolbar,
but it had incorrect selectors (see lines 244-246 of portlet.css). This
rule fixes those selectors. Also, it overrides the rule .aui
.btn-group.open .dropdown-toggle from aui.css which was hiding the icon
when the menu was opened.