Fix function calls by removing renderResponse.getNamespace()
Both exercise 8 and 12 had issues with their function calls to
generateOutput(). renderResponse.getNamespace() was supposed to be used to
get the value of <portlet:namespace />, but it would end up causing an
error that said renderResponse was not defined. Now the function calls
work for onClick in exercise 8 and onSubmit for exercise 12